Fix bugs and reposition the 'receiveTaskResource' function
