+++ host.go
@@ -1,12 +1,18 @@
 package main
 
 import (
+	"bytes"
+	"encoding/binary"
 	"encoding/json"
 	"fmt"
 	"github.com/Equanox/gotron"
+	"io/ioutil"
 	"log"
 	"net"
 	"net/http"
+	"os"
+	"path/filepath"
+	"strings"
 )
 
 type Host struct {
@@ -122,7 +128,7 @@ func (h *Host) receiveWorkerStatus(w http.ResponseWriter, r *http.Request) {
 		log.Fatal("func : receiveWorkerStatus\n", err)
 	}
 
-	h.workers[worker.Id] = &worker
+	h.workers[worker.Id] = &worker // TODO need to create a function for update worker.
 
 	// Update worker status at host's window. ( Method : "Update" )
 	h.send("window.device.status", &worker)
@@ -146,4 +152,40 @@ func (h *Host) receiveTilesInfo(w http.ResponseWriter, r *http.Request) {
 
 	h.tiles = tiles
 	h.DoRender() // tiles : tile information, path : blender file path
+}
+
+func (h *Host) ReceiveTaskResource(w http.ResponseWriter, r *http.Request) {
+	reqBody, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		log.Fatal("func : ReceiveTaskResource\n", err)
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	r.Body = ioutil.NopCloser(bytes.NewBuffer(reqBody))
+
+	fmt.Print("blen : ", string(reqBody[0:4]))
+	if blen := string(reqBody[:4]); strings.Compare(blen, "blen") != 0 {
+		log.Fatal("func : ReceiveTaskResource\n", "This is not a task resource request.")
+	}
+	buffer := bytes.NewBuffer(reqBody[4:8])
+	fileSize, err := binary.ReadVarint(buffer)
+	fmt.Print("size : ", fileSize)
+
+	if err != nil {
+		log.Fatal("func : ReceiveTaskResource\n", "This is not a number.")
+	}
+
+	//if fileSize != len(reqBody[8:]) {
+	//	log.Fatal("func : ReceiveTaskResource\n", "An error occurred while receiving the file.")
+	//}
+
+	dir := filepath.Join(os.TempDir(), "main.blend")
+	err = ioutil.WriteFile(dir, reqBody[8:], 0644)
+	if err != nil {
+		log.Fatal("func : ReceiveTaskResource\n", "File Writing Error.")
+	}
+	fmt.Print("dir:", dir)
+
+	h.filePath <- dir
 }
